Ticker metrics now only updated when needed
Logging tests

Fixing warnings

+++ exchange/test/src/test_utils/macros.hpp
@@ -16,38 +16,38 @@ using SIDE = nutc::messages::SIDE;
 namespace nutc {
 namespace testing_utils {
 bool is_nearly_equal(
-    float f_a, float f_b, float epsilon = std::numeric_limits<float>::epsilon()
+    double f_a, double f_b, double epsilon = std::numeric_limits<double>::epsilon()
 );
 
 bool validate_match(
     const Match& match, const std::string& ticker, const std::string& buyer_id,
-    const std::string& seller_id, messages::SIDE side, float price, float quantity
+    const std::string& seller_id, messages::SIDE side, double price, double quantity
 );
 
 bool validate_ob_update(
-    const ObUpdate& update, const std::string& ticker, messages::SIDE side, float price,
-    float quantity
+    const ObUpdate& update, const std::string& ticker, messages::SIDE side, double price,
+    double quantity
 );
 
 bool validate_market_order(
     const MarketOrder& update, const std::string& client_id, const std::string& ticker,
-    messages::SIDE side, float price, float quantity
+    messages::SIDE side, double price, double quantity
 );
 
 void add_client_simple(
     manager::ClientManager& manager, const std::string& client_id,
-    float capital = STARTING_CAPITAL
+    double capital = STARTING_CAPITAL
 );
 void modify_holdings_simple(
     manager::ClientManager& manager, const std::string& client_id,
-    const std::string& ticker, float quantity
+    const std::string& ticker, double quantity
 );
 
 void modify_capital_simple(
-    manager::ClientManager& manager, const std::string& client_id, float capital_change
+    manager::ClientManager& manager, const std::string& client_id, double capital_change
 );
 
-float get_capital_simple(manager::ClientManager& manager, const std::string& client_id);
+double get_capital_simple(manager::ClientManager& manager, const std::string& client_id);
 
 } // namespace testing_utils
 } // namespace nutc
